LayoutTests/inspector/debugger-step-out.html

<html>
<head>
<script src="../http/tests/inspector/inspector-test.js"></script>
<script src="../http/tests/inspector/debugger-test.js"></script>
<script>

function d() {
    debugger;
}

function testFunction() {
    d();
    notifyDone();
}

function doit()
{
    evaluateInWebInspector("frontend_testStepOut", dumpArray);
}

function frontend_testStepOut(testController)
{
    testController.waitUntilDone();

    WebInspector.showPanel("scripts");
    frontend_ensureDebuggerEnabled(debuggerWasEnabledCallback);

    function debuggerWasEnabledCallback()
    {
        testController.results.push("Debugger was enabled.");
        frontend_evaluateInConsole("setTimeout(testFunction, 0)", function(result)
        {
            testController.results.push("Set timer for test function.");
        });
    }

    frontend_addSniffer(WebInspector, "pausedScript", function(callFrames)
    {
        testController.results.push("Script execution paused.");
        frontend_captureStackTrace(callFrames, testController);
        testStepOut();
    });

    function testStepOut() {
        frontend_addSniffer(WebInspector, "pausedScript", function(callFrames)
        {
            testController.results.push("Script execution paused.");
            frontend_captureStackTrace(callFrames, testController);
            frontend_completeDebuggerTest(testController);
        });
        testController.results.push("Stepping out...");
        document.getElementById("scripts-step-out").click();
    }
}

</script>
</head>

<body onload="onload()">
<p>
Tests "step out" functionality in debugger.
</p>

</body>
</html>
